requirements.txt
torch
torchvision
torchmetrics
opencv-python
Pillow
matplotlib
scikit-image
numpy
tqdm
pandas
openvino
openvino-dev
natsort